+++ build-config/fragments/dev.js
@@ -14,7 +14,6 @@ const contentBase = path.resolve(__dirname, '..', '..');
 
 module.exports = {
     mode: 'development',
-    devtool: 'source-map',
     devServer: {
         contentBase: [contentBase],
         publicPath: '/',