+++ src/js/engine/record-engine.js
@@ -3,6 +3,8 @@
  * @since 2.0.0
  */
 
+import videojs from 'video.js';
+
 import {downloadBlob, addFileInfo} from '../utils/file-util';
 
 const Component = videojs.getComponent('Component');
@@ -57,6 +59,11 @@ class RecordEngine extends Component {
             URL.revokeObjectURL(this.recordedData);
         }
     }
+
+    /**
+     * Destroy engine.
+     */
+    destroy() {}
 
     /**
      * Add filename and timestamp to recorded file object.