+++ src/js/videojs.record.js
@@ -5,6 +5,8 @@
  * MIT license: https://github.com/collab-project/videojs-record/blob/master/LICENSE
  */
 
+import videojs from 'video.js';
+
 import AnimationDisplay from './controls/animation-display';
 import RecordCanvas from './controls/record-canvas';
 import DeviceButton from './controls/device-button';
@@ -19,8 +21,6 @@ import {detectBrowser} from './utils/detect-browser';
 
 import {getAudioEngine, isAudioPluginActive, getConvertEngine, getAnimationEngine} from './engine/engine-loader';
 import {IMAGE_ONLY, AUDIO_ONLY, VIDEO_ONLY, AUDIO_VIDEO, ANIMATION, SCREEN_ONLY, getRecorderMode} from './engine/record-mode';
-
-import videojs from 'video.js';
 
 const Plugin = videojs.getPlugin('plugin');
 const Player = videojs.getComponent('Player');
@@ -78,6 +78,8 @@ class Record extends Plugin {
                 deviceIcon = 'screen-perm';
                 break;
         }
+
+        // add custom interface elements
         DeviceButton.prototype.buildCSSClass = () => {
             // use dynamic icon class
             return 'vjs-record vjs-device-button vjs-control vjs-icon-' + deviceIcon;
@@ -104,9 +106,21 @@ class Record extends Plugin {
         player.cameraButton = new CameraButton(player, options);
         player.cameraButton.hide();
 
-        // add record toggle
+        // add record toggle button
         player.recordToggle = new RecordToggle(player, options);
         player.recordToggle.hide();
+
+        // exclude custom UI elements
+        if (this.player.options_.controlBar) {
+            let customUIElements = ['deviceButton', 'recordIndicator',
+                'cameraButton', 'recordToggle'];
+            customUIElements.forEach((element) => {
+                if (this.player.options_.controlBar[element] !== undefined) {
+                    this.player[element].layoutExclude = true;
+                    this.player[element].hide();
+                }
+            });
+        }
 
         // wait until player ui is ready
         this.player.one('ready', this.setupUI.bind(this));
@@ -222,8 +236,10 @@ class Record extends Plugin {
                 this.player.removeTechControlsListeners_();
 
                 if (this.player.options_.controls) {
-                    // progress control isn't used by this plugin
-                    this.player.controlBar.progressControl.hide();
+                    // progress control isn't used by this plugin, hide if present
+                    if (this.player.controlBar.progressControl !== undefined) {
+                        this.player.controlBar.progressControl.hide();
+                    }
 
                     // prevent controlbar fadeout
                     this.player.on('userinactive', (event) => {
@@ -248,8 +264,10 @@ class Record extends Plugin {
         // display max record time
         this.setDuration(this.maxLength);
 
-        // hide play control
-        this.player.controlBar.playToggle.hide();
+        // hide play control (if present)
+        if (this.player.controlBar.playToggle !== undefined) {
+            this.player.controlBar.playToggle.hide();
+        }
     }
 
     /**
@@ -436,7 +454,9 @@ class Record extends Plugin {
         this.setCurrentTime(0);
 
         // hide play/pause control (e.g. when stopDevice was used)
-        this.player.controlBar.playToggle.hide();
+        if (this.player.controlBar.playToggle !== undefined) {
+            this.player.controlBar.playToggle.hide();
+        }
 
         // reset playback listeners
         this.off(this.player, 'timeupdate', this.playbackTimeUpdate);
@@ -533,12 +553,9 @@ class Record extends Plugin {
 
             // show elements that should never be hidden in animation,
             // audio and/or video modus
-            let uiElements = [
-                this.player.controlBar.currentTimeDisplay,
-                this.player.controlBar.timeDivider,
-                this.player.controlBar.durationDisplay
-            ];
+            let uiElements = ['currentTimeDisplay', 'timeDivider', 'durationDisplay'];
             uiElements.forEach((element) => {
+                element = this.player.controlBar[element];
                 if (element !== undefined) {
                     element.el().style.display = 'block';
                     element.show();
@@ -616,7 +633,9 @@ class Record extends Plugin {
             this._recording = true;
 
             // hide play/pause control
-            this.player.controlBar.playToggle.hide();
+            if (this.player.controlBar.playToggle !== undefined) {
+                this.player.controlBar.playToggle.hide();
+            }
 
             // reset playback listeners
             this.off(this.player, 'timeupdate', this.playbackTimeUpdate);
@@ -820,8 +839,10 @@ class Record extends Plugin {
         this.player.recordedData = this.engine.recordedData;
 
         // change the replay button back to a play button
-        this.player.controlBar.playToggle.removeClass('vjs-ended');
-        this.player.controlBar.playToggle.show();
+        if (this.player.controlBar.playToggle !== undefined) {
+            this.player.controlBar.playToggle.removeClass('vjs-ended');
+            this.player.controlBar.playToggle.show();
+        }
 
         // notify converter
         if (this.converter !== undefined) {
@@ -1113,6 +1134,7 @@ class Record extends Plugin {
         // prevent callbacks if recording is in progress
         if (this.engine) {
             this.engine.dispose();
+            this.engine.destroy();
             this.engine.off('recordComplete', this.engineStopCallback);
         }
 
@@ -1190,7 +1212,9 @@ class Record extends Plugin {
         }
 
         // hide play control
-        this.player.controlBar.playToggle.hide();
+        if (this.player.controlBar.playToggle !== undefined) {
+            this.player.controlBar.playToggle.hide();
+        }
 
         // show device selection button
         this.player.deviceButton.show();
@@ -1458,7 +1482,7 @@ class Record extends Plugin {
             return;
         }
 
-        // List cameras and microphones.
+        // list video and audio devices
         navigator.mediaDevices.enumerateDevices(this).then((devices) => {
             this.devices = [];
             devices.forEach((device) => {
@@ -1545,6 +1569,4 @@ if (videojs.getPlugin('record') === undefined) {
 }
 
 // export plugin
-module.exports = {
-    Record
-};
+export {Record};